+++ replacer-backend/src/main/java/es/bvalero/replacer/page/PageReviewService.java
@@ -19,6 +19,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.VisibleForTesting;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -107,6 +108,9 @@ abstract class PageReviewService {
         String key = buildReplacementCacheKey(options);
         PageSearchResult result = cachedPageIds.getIfPresent(key);
         if (result != null && result.isEmptyTotal()) {
+            if (StringUtils.isNotBlank(options.getType()) && StringUtils.isNotBlank(options.getSubtype())) {
+                replacementService.reviewAsSystemBySubtype(options.getLang(), options.getType(), options.getSubtype());
+            }
             return false;
         } else {
             PageSearchResult pageIds = findPageIdsToReview(options);