+++ replacer-backend/src/main/java/es/bvalero/replacer/replacement/ReplacementStatsDao.java
@@ -3,6 +3,7 @@ package es.bvalero.replacer.replacement;
 import es.bvalero.replacer.common.ReplacerException;
 import es.bvalero.replacer.common.WikipediaLanguage;
 import java.util.List;
+import org.springframework.lang.Nullable;
 
 interface ReplacementStatsDao {
     long countReplacementsReviewed(WikipediaLanguage lang);
@@ -12,6 +13,14 @@ interface ReplacementStatsDao {
     List<ReviewerCount> countReplacementsGroupedByReviewer(WikipediaLanguage lang);
 
     LanguageCount countReplacementsGroupedByType(WikipediaLanguage lang) throws ReplacerException;
+
+    void reviewByPageId(
+        WikipediaLanguage lang,
+        int pageId,
+        @Nullable String type,
+        @Nullable String subtype,
+        String reviewer
+    );
 
     void reviewAsSystemBySubtype(WikipediaLanguage lang, String type, String subtype);
 }