+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Versions
 
+## 2.9.1 (28 Feb 2021)
+- Fix: load replacement list every time the list section is opened
+- Fix: maintain replacement count cache in backend
+
 ## 2.9.0 (27 Feb 2021)
 - Improve texts in dashboard, removing references to Spanish language and adding a warning.
 - Improve suggestions for composed misspellings